디버깅중

+++ Project1/Project1/src/Neighborhood.h
@@ -8,7 +8,7 @@ namespace ms
 {
 struct Neighbor_Informations
 {
-  std::vector<size_t>  indexes;
+  std::vector<UINT>  indexes;
   std::vector<Vector3> translate_vectors;
   std::vector<float>   distances;
 };
@@ -23,7 +23,7 @@ public:
     const std::vector<Vector3>& fluid_particle_pos_vectors,
     const std::vector<Vector3>& boundary_particle_pos_vectors) = 0;
 
-  virtual const Neighbor_Informations& search_for_fluid(const size_t fpid) const    = 0;
-  virtual const std::vector<size_t>&   search_for_boundary(const size_t bpid) const = 0;
+  virtual const Neighbor_Informations& search_for_fluid(const UINT fpid) const    = 0;
+  virtual const std::vector<UINT>&   search_for_boundary(const UINT bpid) const = 0;
 };
 } // namespace ms
